+++ src/helper/dir.rs
@@ -1,0 +1,23 @@
+use std::fs;
+use std::io;
+use std::path::Path;
+
+// Function to copy a dir recursivly
+pub fn copy_dir_recursive(from: &Path, to: &Path) -> io::Result<()> {
+    if !to.exists() {
+        fs::create_dir_all(to)?;
+    }
+
+    for entry_result in fs::read_dir(from)? {
+        let entry = entry_result?;
+        let from_path = entry.path();
+        let to_path = to.join(entry.file_name());
+
+        if from_path.is_dir() {
+            copy_dir_recursive(&from_path, &to_path)?;
+        } else {
+            fs::copy(&from_path, &to_path)?;
+        }
+    }
+    Ok(())
+}