+++ src/helper/mod.rs
@@ -1,3 +1,4 @@
+pub mod dir;
 pub mod macros;
 pub mod print;
 pub mod update;